# DOCS/api/openapi.yaml
openapi: 3.1.0
info:
  title: CRM FreeLux API
  description: |
    API REST para el sistema CRM de Aceros y Perfiles FreeLux.

    ## Autenticación
    La API usa JWT Bearer tokens. Obtén un token en `POST /api/v1/auth/login`
    y envíalo en el header `Authorization: Bearer <token>`.

    ## Roles
    | Rol | Descripción |
    |-----|-------------|
    | `ADMIN` | Acceso total al sistema |
    | `GERENTE` | Aprobación de precios, reportes, gestión de usuarios |
    | `VENDEDOR` | Crear cotizaciones, ver clientes y productos |
    | `COMPRAS` | Gestión de proveedores, órdenes de compra |
    | `ALMACEN` | Control de inventario, entradas/salidas |

  version: 1.0.0
  contact:
    name: FreeLux CRM
    email: dev@freelux.mx
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Desarrollo local
  - url: https://crm.freelux.mx
    description: Staging / Producción
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Token JWT obtenido de POST /api/v1/auth/login"
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "admin@freelux.mx"
        password:
          type: string
          minLength: 8
          example: "Admin1234!"
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          description: "Segundos hasta expiración del access_token"
          example: 1800

    # ── Users ──────────────────────────────────────────────────
...
    ClientCreate:
      type: object
      required: [company_name, rfc]
      properties:
        company_name:
          type: string
          example: "IIMSA Infraestructura Industrial SA de CV"
        rfc:
          type: string
          pattern: '^[A-ZÑ&]{3,4}[0-9]{6}[A-Z0-9]{3}$'
          example: "III880415AB7"
        contact_name:
          type: string
          example: "Ing. Roberto Flores"
        phone:
          type: string
          example: "+52 55 1234 5678"
        email:
          type: string
          format: email
          example: "compras@iimsa.com.mx"
        address:
          type: string
          example: "Blvd. Manuel Ávila Camacho 32, Polanco, CDMX"
        credit_limit:
          type: number
          format: float
          example: 500000.00
    ClientResponse:
      allOf:
        - $ref: '#/components/schemas/ClientCreate'
        - type: object
          properties:
            id: { type: integer, example: 42 }
            created_at: { type: string, format: date-time }
            total_purchased_ytd: { type: number, example: 284500.00 }
            last_purchase_date: { type: string, format: date }
            quotes_count: { type: integer, example: 12 }

    # ── Products ───────────────────────────────────────────────
    ProductCreate:
      type: object
      required: [sku, name, unit, cost_price, list_price]
      properties:
        sku:
          type: string
          example: "PTR-4X4-14-6M"
        name:
          type: string
          example: "PTR 4\"x4\"x6m Calibre 14"
        description:
          type: string
          example: "Tubo estructural rectangular 4x4 pulgadas, espesor calibre 14 (1.9mm), longitud 6 metros, acero A36"
        category:
          type: string
          enum: [PERFIL, LAMINA, TUBERIA, PTR, HERRAJE, MALLA, VARILLA, OTRO]
          example: PTR
        unit:
          type: string
          example: "PZA"
        cost_price:
          type: number
          format: float
          example: 820.00
        list_price:
          type: number
          format: float
          example: 980.00
        stock_min:
          type: integer
          example: 20
        stock_max:
          type: integer
          example: 200
        weight_kg:
          type: number
          example: 38.5
        fulfillment_type:
          type: string
          enum: [STOCK, MTO]
          example: STOCK
          description: "STOCK = disponible en almacén, MTO = Make To Order (se pide al recibir venta)"
    ProductResponse:
      allOf:
        - $ref: '#/components/schemas/ProductCreate'
        - type: object
          properties:
            id: { type: integer, example: 15 }
            current_stock: { type: integer, example: 45 }
            stock_status:
              type: string
              enum: [OK, LOW, OUT]
              example: LOW
            created_at: { type: string, format: date-time }

    # ── Suppliers ──────────────────────────────────────────────
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
            example:
              email: admin@freelux.mx
              password: Admin1234!
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              example: { detail: "Credenciales inválidas" }
        '422':
          description: Error de validación
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Renovar access token
      security: []
      requestBody:
        content:
          application/json:
            example:
              refresh_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Token renovado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401':
          description: Refresh token inválido o expirado
...
  /api/v1/clients:
    get:
      tags: [Clients]
      summary: Listar clientes
      parameters:
        - name: q
          in: query
          description: Búsqueda por nombre, RFC o email
          schema: { type: string, example: "IIMSA" }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              example:
                items:
                  - id: 42
                    company_name: "IIMSA Infraestructura Industrial SA de CV"
                    rfc: "III880415AB7"
                    contact_name: "Ing. Roberto Flores"
                    phone: "+52 55 1234 5678"
                    credit_limit: 500000.00
                    total_purchased_ytd: 284500.00
                meta:
                  total: 47
                  page: 1
                  per_page: 20
    post:
      tags: [Clients]
      summary: Crear cliente
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClientCreate' }
      responses:
        '201':
          description: Cliente creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClientResponse' }
  /api/v1/clients/{client_id}:
    get:
      tags: [Clients]
      summary: Obtener cliente
      parameters:
        - name: client_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClientResponse' }
    put:
      tags: [Clients]
      summary: Actualizar cliente
      parameters:
        - name: client_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClientCreate' }
      responses:
        '200':
          description: Cliente actualizado
...
  /api/v1/products:
    get:
      tags: [Products]
      summary: Listar productos
      parameters:
        - name: category
          in: query
          schema: { type: string, example: "PTR" }
        - name: stock_status
          in: query
          schema: { type: string, enum: [OK, LOW, OUT] }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              example:
                items:
                  - id: 15
                    sku: "PTR-4X4-14-6M"
                    name: "PTR 4\"x4\"x6m Calibre 14"
                    category: "PTR"
                    unit: "PZA"
                    cost_price: 820.00
                    list_price: 980.00
                    current_stock: 12
                    stock_min: 20
                    stock_status: "LOW"
    post:
      tags: [Products]
      summary: Crear producto
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductCreate' }
      responses:
        '201':
          description: Producto creado
...
  /api/v1/products/{product_id}:
    get:
      tags: [Products]
      summary: Obtener producto
      parameters:
        - name: product_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductResponse' }
    put:
      tags: [Products]
      summary: Actualizar producto
      parameters:
        - name: product_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductCreate' }
      responses:
        '200':
          description: Producto actualizado

  # ── Suppliers ──────────────────────────────────────────────
...